--- save-backend/src/main/kotlin/com/saveourtool/save/backend/utils/ConvertingAuthenticationManager.kt
@@ -1,68 +1,0 @@
-//package com.saveourtool.save.backend.utils
-//
-//import com.saveourtool.save.backend.service.UserDetailsService
-//import com.saveourtool.save.authservice.utils.AuthenticationDetails
-//import com.saveourtool.save.authservice.utils.IdentitySourceAwareUserDetails
-//import org.springframework.beans.factory.annotation.Autowired
-//import org.springframework.security.authentication.BadCredentialsException
-//import org.springframework.security.authentication.ReactiveAuthenticationManager
-//import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-//import org.springframework.security.core.Authentication
-//import org.springframework.stereotype.Component
-//import reactor.core.publisher.Mono
-//import reactor.kotlin.core.publisher.cast
-//import reactor.kotlin.core.publisher.switchIfEmpty
-//
-///**
-// * Implementation of `ReactiveAuthenticationManager` that doesn't check user credentials
-// * and simply authenticates every user with username. Should be used in secure environment,
-// * where user identity is already guaranteed.
-// */
-//@Component
-//class ConvertingAuthenticationManager(
-//    @Autowired private var userDetailsService: UserDetailsService
-//) : ReactiveAuthenticationManager {
-//    /**
-//     * Authenticate user, by checking the received data, which converted into UsernamePasswordAuthenticationToken
-//     * by [CustomAuthenticationBasicConverter] with record in DB
-//     *
-//     * @return augmented mono of UsernamePasswordAuthenticationToken with additional details
-//     * @throws BadCredentialsException in case of bad credentials
-//     */
-//    override fun authenticate(authentication: Authentication): Mono<Authentication> = if (authentication is UsernamePasswordAuthenticationToken) {
-//        val identitySource = (authentication.details as AuthenticationDetails).identitySource
-//        if (identitySource == null || !authentication.name.startsWith("$identitySource:")) {
-//            throw BadCredentialsException(authentication.name)
-//        }
-//        val name = authentication.name.drop(identitySource.length + 1)
-//        userDetailsService.findByUsername(name)
-//            .cast<IdentitySourceAwareUserDetails>()
-//            .filter { it.identitySource == identitySource }
-//            .switchIfEmpty {
-//                Mono.error { BadCredentialsException(name) }
-//            }
-//            .onErrorMap {
-//                BadCredentialsException(name)
-//            }
-//            .map {
-//                it.toAuthenticationWithDetails(authentication)
-//            }
-//    } else {
-//        Mono.error { BadCredentialsException("Unsupported authentication type ${authentication::class}") }
-//    }
-//
-//    /**
-//     * Fixme: since identitySource is in `AuthenticationDetails`, it can be removed from principal
-//     */
-//    private fun IdentitySourceAwareUserDetails.toAuthenticationWithDetails(authentication: Authentication) =
-//            UsernamePasswordAuthenticationToken(
-//                "$identitySource:$username",
-//                authentication.credentials,
-//                authorities
-//            ).apply {
-//                details = AuthenticationDetails(
-//                    id = id,
-//                    identitySource = identitySource,
-//                )
-//            }
-//}
