+++ save-backend/src/test/kotlin/com/saveourtool/save/backend/security/BasicSecurityTest.kt
@@ -1,10 +1,10 @@
 package com.saveourtool.save.backend.security
 
+import com.saveourtool.save.authservice.security.ConvertingAuthenticationManager
 import com.saveourtool.save.backend.repository.OriginalLoginRepository
 import com.saveourtool.save.backend.repository.UserRepository
 import com.saveourtool.save.backend.service.UserDetailsService
 import com.saveourtool.save.authservice.utils.AuthenticationDetails
-import com.saveourtool.save.backend.utils.ConvertingAuthenticationManager
 import com.saveourtool.save.entities.User
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
